+++ testing/testing_parse_opts.cpp
@@ -11,22 +11,24 @@ int main( int argc, char** argv )
     opts.parse_opts( argc, argv );
     
     for( int itest = 0; itest < opts.ntest; ++itest ) {
-        printf( "m %5ld, n %5ld, k %5ld\n",
-                long(opts.msize[itest]), long(opts.nsize[itest]), long(opts.ksize[itest]) );
+        printf( "m %5lld, n %5lld, k %5lld\n",
+                (long long) opts.msize[itest],
+                (long long) opts.nsize[itest],
+                (long long) opts.ksize[itest] );
     }
     printf( "\n" );
     
-    printf( "ntest    %ld\n", long(opts.ntest) );
+    printf( "ntest    %lld\n", (long long) opts.ntest );
     printf( "\n" );
     
-    printf( "nb       %ld\n", long(opts.nb)       );
-    printf( "nrhs     %ld\n", long(opts.nrhs)     );
-    printf( "nqueue   %ld\n", long(opts.nqueue)   );
-    printf( "ngpu     %ld\n", long(opts.ngpu)     );
-    printf( "niter    %ld\n", long(opts.niter)    );
-    printf( "nthread  %ld\n", long(opts.nthread)  );
-    printf( "itype    %ld\n", long(opts.itype)    );
-    printf( "verbose  %ld\n", long(opts.verbose)  );
+    printf( "nb       %lld\n", (long long) opts.nb       );
+    printf( "nrhs     %lld\n", (long long) opts.nrhs     );
+    printf( "nqueue   %lld\n", (long long) opts.nqueue   );
+    printf( "ngpu     %lld\n", (long long) opts.ngpu     );
+    printf( "niter    %lld\n", (long long) opts.niter    );
+    printf( "nthread  %lld\n", (long long) opts.nthread  );
+    printf( "itype    %lld\n", (long long) opts.itype    );
+    printf( "verbose  %lld\n", (long long) opts.verbose  );
     printf( "\n" );
     
     printf( "check    %s\n", (opts.check  ? "true" : "false") );
@@ -44,7 +46,7 @@ int main( int argc, char** argv )
     printf( "jobvl    %3d (%s)\n", opts.jobvl,  lapack_vec_const(   opts.jobvl  ));
     
     for( auto iter = opts.svd_work.begin(); iter < opts.svd_work.end(); ++iter ) {
-        printf( "svd_work %ld\n", long(*iter) );
+        printf( "svd_work %lld\n", (long long) *iter );
     }
     for( auto iter = opts.jobu.begin(); iter < opts.jobu.end(); ++iter ) {
         printf( "jobu     %3d (%s)\n", *iter, lapack_vec_const( *iter ));
